fix loadfile error on mpv master
If we don't use the options argument the loadfile command should work
on all versions past 0.28.1, which is now not in any supported
distribution. We reset all our settings when we play new files so this
should work just fine.

partial revert of cd262365776fc0cf8588746d003abdea1c0323d9